requirements.txt
torch==2.6.0
torchvision==0.21.0
transformers==4.55.4
accelerate==1.7.0
safetensors==0.5.3

timm==1.0.16
opencv-python-headless==4.11.0.86
pillow==11.0.0
qwen-vl-utils==0.0.11
decord==0.6.0

peft==0.14.0

numpy==1.26.4
pandas==2.2.3
datasets==3.6.0
pyarrow==20.0.0

tqdm==4.67.1
einops==0.8.1
huggingface-hub==0.34.4
sentencepiece==0.2.0
tokenizers==0.21.1

fastapi==0.115.12
uvicorn==0.34.2
starlette==0.46.2

outlines==0.1.11
langchain==0.3.27
openai==1.97.1